Conflicts:
	audio_alsa.c
	audio_pulse.c
	configure
	player.c
	rtp.c
	rtsp.c

+++ audio_alsa.c
@@ -211,6 +211,6 @@ static long long get_delay(void) {
     frames = 0;
   }
 
-  return (long long)(frames * (1000000L / device_sample_rate));
+  return (long long)frames * 1000000LL / (long long)device_sample_rate;
 }
